third_party/blink/web_tests/platform/generic/external/wpt/service-workers/service-worker/unregister-then-register.https-expected.txt

This is a testharness.js-based test.
PASS Unregister then register resolves to a new value
FAIL Unregister then register does not resolve to the original value even if the registration is in use. assert_not_equals: Unregister and register should always create a new registration got disallowed value object "[object ServiceWorkerRegistration]"
PASS Unregister then register does not affect existing controllee
FAIL Unregister then register does not resurrect the registration assert_equals: Registration is new expected null but got object "[object ServiceWorker]"
Harness: the test ran to completion.
